moved files from /continent/ to /zones/; separated Actor for managing items on the ground; added comments
